adding back

+++ blog.py
@@ -9,6 +9,7 @@ import jinja2
 from user import User
 from post import Post
 from like import Like
+from comment import Comment
 from google.appengine.ext import db
 
 template_dir = os.path.join(os.path.dirname(__file__), 'templates')
@@ -112,6 +113,9 @@ class PostPage(BlogHandler):
         key = db.Key.from_path('Post', int(post_id), parent=blog_key())
         post = db.get(key)
 
+        comments = db.GqlQuery("select * from Comment where post_id = " +
+                               post_id + " order by created desc")
+
         likes = db.GqlQuery("select * from Like where post_id="+post_id)
 
         if not post:
@@ -121,7 +125,7 @@ class PostPage(BlogHandler):
         error = self.request.get('error')
 
         self.render("permalink.html", post=post, noOfLikes=likes,
-                    error=error)
+                    comments=comments,error=error)
 
     def post(self, post_id):
         key = db.Key.from_path('Post', int(post_id), parent=blog_key())
@@ -130,6 +134,12 @@ class PostPage(BlogHandler):
         if not post:
             self.error(404)
             return
+
+        """
+            On posting comment, new comment tuple is created and stored,
+            with relationship data of user and post.
+        """
+        c = ""
 
         if(self.user):
             # On clicking like, post-like value increases.
@@ -149,8 +159,26 @@ class PostPage(BlogHandler):
                              post_id=int(post_id))
                     l.put()
 
+             # On commenting, it creates new comment tuple
+            if(self.request.get('comment')):
+                c = Comment(parent=blog_key(), user_id=self.user.key().id(),
+                            post_id=int(post_id),
+                            comment=self.request.get('comment'))
+                c.put()
+        else:
+            self.redirect("/login?error=You need to login before " +
+                            "performing edit, like or commenting.!!")
+            return
+
+        comments = db.GqlQuery("select * from Comment where post_id = " +
+                                post_id + "order by created desc")
+
+        likes = db.GqlQuery("select * from Like where post_id="+post_id)
+
         self.render("permalink.html", post=post,
-                    noOfLikes=likes.count())
+                    comments=comments,
+                    noOfLikes=likes.count(),
+                    new=c)
 
 
 class NewPost(BlogHandler):
@@ -187,6 +215,8 @@ class EditPost(BlogHandler):
         if self.user:
             key = db.Key.from_path('Post', int(post_id), parent=blog_key())
             post = db.get(key)
+            if not post:
+                return self.redirect('/login')
             if post.user_id == self.user.key().id():
                 self.render("editpost.html", subject=post.subject,
                             content=post.content)
@@ -208,6 +238,11 @@ class EditPost(BlogHandler):
         if self.user:
             key = db.Key.from_path('Post', int(post_id), parent=blog_key())
             post = db.get(key)
+
+            if not post:
+                self.error(404)
+                return
+
             if post.user_id == self.user.key().id():
                 self.render("editpost.html", subject=post.subject,
                             content=post.content)
@@ -239,6 +274,11 @@ class DeletePost(BlogHandler):
         if self.user:
             key = db.Key.from_path('Post', int(post_id), parent=blog_key())
             post = db.get(key)
+
+            if not post:
+                self.error(404)
+                return
+
             if post.user_id == self.user.key().id():
                 post.delete()
                 self.redirect("/?deleted_post_id="+post_id)
@@ -248,6 +288,93 @@ class DeletePost(BlogHandler):
         else:
             self.redirect("/login?error=You need to be logged, in order" +
                           " to delete your post!!")
+
+class DeleteComment(BlogHandler):
+    def get(self, post_id, comment_id):
+        if self.user:
+            key = db.Key.from_path('Comment', int(comment_id),
+                                   parent=blog_key())
+            c = db.get(key)
+            key1 = db.Key.from_path('Post', int(post_id), parent=blog_key())
+            post = db.get(key1)
+
+            if not post:
+                self.error(404)
+                return
+
+            if not c:
+                self.error(404)
+                return
+
+            if c.user_id == self.user.key().id():
+                c.delete()
+                self.redirect("/blog/"+post_id+"?deleted_comment_id=" +
+                              comment_id)
+            else:
+                self.redirect("/blog/" + post_id + "?error=You don't have " +
+                                "access to delete this comment.")
+        else:
+            self.redirect("/login?error=You need to be logged, in order to " +
+                        "delete your comment!!")
+
+class EditComment(BlogHandler):
+    def get(self, post_id, comment_id):
+        if self.user:
+            key = db.Key.from_path('Comment', int(comment_id),
+                                   parent=blog_key())
+            c = db.get(key)
+            key1 = db.Key.from_path('Post', int(post_id), parent=blog_key())
+            post = db.get(key1)
+
+            if not post:
+                self.error(404)
+                return
+
+            if not c:
+                self.error(404)
+                return
+
+            if c.user_id == self.user.key().id():
+                self.render("editcomment.html", comment=c.comment)
+            else:
+                self.redirect("/blog/" + post_id +
+                            "?error=You don't have access to edit this " +
+                            "comment.")
+        else:
+            self.redirect("/login?error=You need to be logged, in order to" +
+                        " edit your post!!")
+
+    def post(self, post_id, comment_id):
+        """
+            Updates post.
+        """
+        if not self.user:
+            self.redirect('/login')
+
+        key1 = db.Key.from_path('Post', int(post_id), parent=blog_key())
+        post = db.get(key1)
+
+        if not post:
+            self.error(404)
+            return
+
+        comment = self.request.get('comment')
+
+        if comment:
+            key = db.Key.from_path('Comment',
+                                    int(comment_id), parent=blog_key())
+            c = db.get(key)
+            if not c:
+                self.error(404)
+                return
+                
+            c.comment = comment
+            c.put()
+            self.redirect('/blog/%s' % post_id)
+        else:
+            error = "subject and content, please!"
+            self.render("editpost.html", subject=subject,
+                        content=content, error=error)
 
 
 USER_RE = re.compile(r"^[a-zA-Z0-9_-]{3,20}$")
@@ -352,6 +479,10 @@ app = webapp2.WSGIApplication([('/?', BlogFront),
                                ('/blog/newpost', NewPost),
                                ('/blog/deletepost/([0-9]+)', DeletePost),
                                ('/blog/editpost/([0-9]+)', EditPost),
+                               ('/blog/deletecomment/([0-9]+)/([0-9]+)',
+                                DeleteComment),
+                               ('/blog/editcomment/([0-9]+)/([0-9]+)',
+                                EditComment),
                                ('/signup', Register),
                                ('/login', Login),
                                ('/logout', Logout),
